Hanoi Tower Recursive
Hanoi tower is a way arrange disks in correct order. Recursive means idea of breakdown the problem to the similar parts to sub problems. it means we move the the top n-1 disks to the auxiliary peg, then the largest disk to the destination, and finally the n-1 disks from the auxiliary to the destination.